execution/requirements.txt
# Python dependencies for execution scripts
# Install with: pip install -r requirements.txt

# Core dependencies
python-dotenv>=1.0.0
requests>=2.31.0

# Web scraping dependencies
selenium>=4.15.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Google Sheets API dependencies
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.110.0
